refactor(useFetchOrder): Bill 로직 분리
